+++ skfem/experimental/__init__.py
@@ -1,0 +1,6 @@
+import logging
+
+
+logger = logging.getLogger(__name__)
+logger.warning("You are using experimental features under skfem.experimental "
+               "that might change at any time without prior notice.")
